requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
insightface>=0.7.3
onnxruntime>=1.16.0
faiss-cpu>=1.7.4
ultralytics>=8.3.0
mediapipe==0.10.14
opencv-python>=4.9.0
numpy>=1.24.0
scipy>=1.13.0
Pillow>=10.0.0
pandas>=2.2.2
tqdm>=4.66.4
python-dotenv>=1.0.0
requests>=2.31.0
